Tests/LibWeb/Text/expected/Fetch/fetch-later-activate-after.txt

activated before delay elapsed: false
activated after timeout: true
cookie observed: true
PASS
